Fix and restructure all tests for changed frame API
