+++ tox.ini
@@ -5,10 +5,12 @@ envlist = py
 deps=pytest
 # setenv = PYTHONMALLOC = pymalloc
 #     PYTHONMALLOCSTATS = 'yes'
-commands=pytest tests/block/test_block.py
+commands=
+    pytest tests/block/test_block.py
     pytest --tb=long --showlocals tests/frame/test_frame_0.py
     pytest --tb=long --showlocals tests/frame/test_frame_1.py
     pytest --tb=long --showlocals tests/frame/test_frame_2.py
     pytest --tb=long --showlocals tests/frame/test_frame_3.py
     pytest --tb=long --showlocals tests/frame/test_frame_4.py
     pytest --tb=long --showlocals tests/frame/test_frame_5.py
+    pytest --tb=long --showlocals tests/frame/test_frame_6.py
